+++ tools/src/main.cpp
@@ -5,13 +5,14 @@ auto main(int argc, char** argv) -> int {
 	using namespace bave::tools;
 
 	auto const args = bave::MainArgs{argc, argv};
-	auto const daci = bave::DesktopApp::CreateInfo{
+	auto assets_path = args.upfind_assets_dir("assets,example/assets");
+	auto daci = bave::DesktopApp::CreateInfo{
 		.args = args,
 		.title = "Bave Tools",
 		.mode = bave::Windowed{.extent = Applet::window_size_v, .lock_aspect_ratio = false},
-		.assets_dir = args.find_assets_super_dir("assets,example/assets"),
+		.data_loader = std::make_unique<bave::FileLoader>(assets_path),
 	};
-	auto app = bave::DesktopApp{daci};
-	app.set_bootloader(Bootloader{});
+	auto app = bave::DesktopApp{std::move(daci)};
+	app.set_bootloader(Bootloader{std::move(assets_path)});
 	return static_cast<int>(app.run());
 }